完善职工合同管理，部门组织架构管理，职位管理。

+++ src/main/java/com/lqs/hrm/controller/DepartmentLevelController.java
@@ -22,12 +22,16 @@ import com.github.pagehelper.PageInfo;
 import com.github.pagehelper.util.StringUtil;
 import com.lqs.hrm.entity.Department;
 import com.lqs.hrm.entity.DepartmentLevel;
+import com.lqs.hrm.entity.Employee;
+import com.lqs.hrm.entity.EmployeePosition;
 import com.lqs.hrm.entity.User;
 import com.lqs.hrm.json.JsonCommonResult;
 import com.lqs.hrm.json.JsonPageResult;
 import com.lqs.hrm.service.impl.DepartmentLevelServiceImpl;
 import com.lqs.hrm.service.impl.DepartmentServiceImpl;
+import com.lqs.hrm.service.impl.EmployeePositionServiceImpl;
 import com.lqs.hrm.service.impl.EmployeeServiceImpl;
+import com.lqs.hrm.service.impl.PositionServiceImpl;
 import com.lqs.hrm.service.impl.StatusServiceImpl;
 import com.lqs.hrm.util.PageRequest;
 import com.lqs.hrm.util.PageResult;
@@ -44,6 +48,10 @@ public class DepartmentLevelController {
 	private EmployeeServiceImpl employeeService;
 	@Autowired
 	private StatusServiceImpl statusService;
+	@Autowired
+	private EmployeePositionServiceImpl employeePositionService;
+	@Autowired
+	private PositionServiceImpl positionService;
 	
 	/**
 	 * 查询部门并跳转至部门架构管理页面
@@ -175,9 +183,24 @@ public class DepartmentLevelController {
 			for (int i = 0; i < list.size(); i++) {
 				//设置部门级别
 				list.get(i).setDlLeve(departmentLevelService.get(list.get(i).getDlId()).getLevel());
-				//设置部门主管名称
-				if(list.get(i).getManageEmpjobid() != null && !list.get(i).getManageEmpjobid().isEmpty()) {
-					list.get(i).setManageEmpName(employeeService.get(list.get(i).getManageEmpjobid()).getEmpName());
+				//设置部门主管职位名称
+				if (list.get(i).getManagePositionid() != null && list.get(i).getManagePositionid().intValue()!= 0) {
+					list.get(i).setManagePositionName(positionService.get(list.get(i).getManagePositionid()).getPositionName());
+				}
+				//设置部门主管人工号和姓名
+				if(list.get(i).getManagePositionid() != null && list.get(i).getManagePositionid() != 0) {
+					//获取部门主管职位
+					List<EmployeePosition> employeePositionList = employeePositionService.listByPositionId(list.get(i).getManagePositionid());
+					if (employeePositionList == null || employeePositionList.size() == 0) {
+						//该部门主管职位还未分配给职工
+						list.get(i).setManageEmpName("");
+					}else {
+						//该部门主管职位还已分配给职工，则查找该职工信息
+						Employee employee = employeeService.get(employeePositionList.get(0).getEmpJobid());
+						//设置部门主管人工号和姓名
+						list.get(i).setManageEmpJobId(employee.getEmpJobid());
+						list.get(i).setManageEmpName(employee.getEmpName());
+					}
 				}
 				//设置上级部门名称
 				if (list.get(i).getParentId() != null) {
@@ -316,6 +339,12 @@ public class DepartmentLevelController {
 	public JsonCommonResult<Object> delete(HttpServletRequest request) {
 		//获取级别Id 
 		String dlIdStr = request.getParameter("dlId");		
+		//检查该部门级别下面是否还有部门信息
+		List<Department> departmentList = departmentService.listByDlId(Integer.valueOf(dlIdStr));
+		if (departmentList != null && departmentList.size() != 0) {
+			//该部门级别下面还有部门信息，不能删除
+			return new JsonCommonResult<>("100", null, "该部门级别下面还有部门信息，不能删除！");
+		}
 		int result = departmentLevelService.delete(Integer.valueOf(dlIdStr));
 		if (result == 0) {
 			return new JsonCommonResult<>("100", null, "删除失败");
